adding invoice and working with files using PDFDocument

+++ controllers/admin.js
@@ -1,4 +1,6 @@
 const mongoose = require("mongoose");
+const fileHelper = require('../util/file');
+
 const {
     validationResult
 } = require("express-validator/check");
@@ -147,6 +149,7 @@ exports.postEditProduct = (req, res, next) => {
             product.description = updatedDescription;
 
             if (image) {
+                fileHelper.deleteFile(product.imageUrl);
                 product.imageUrl = image.path;
             }
 
@@ -185,9 +188,16 @@ exports.getProducts = (req, res, next) => {
 exports.postDeleteProduct = (req, res, next) => {
     const prodId = req.body.productId;
 
-    Product.deleteOne({
-            _id: prodId,
-            userId: req.user._id
+    Product.findById(prodId)
+        .then(product => {
+            if (!product) {
+                return next(new Error('Product not found'));
+            }
+            fileHelper.deleteFile(product.imageUrl);
+            return Product.deleteOne({
+                _id: prodId,
+                userId: req.user._id
+            })
         })
         .then(result => {
             console.log(`Product Deleted`);
